+++ lib/plugins/aws/deployFunction.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const crypto = require('crypto');
 const path = require('path');
 const fs = require('fs');
-const validate = require('../lib/validate');
+const validate = require('./lib/validate');
 const filesize = require('filesize');
 
 class AwsDeployFunction {
