+++ lib/plugins/index.js
@@ -52,7 +52,7 @@ module.exports = [
   require('./aws/package/compile/events/sqs/index.js'),
   require('./aws/package/compile/events/cloudFront/index.js'),
   require('./aws/package/compile/events/httpApi'),
-  require('./aws/deployFunction/index.js'),
+  require('./aws/deployFunction.js'),
   require('./aws/deployList/index.js'),
   require('./aws/invokeLocal/index.js'),
   require('./standalone'),
